% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.acrophase.msf.R
\name{correct.acrophase.msf}
\alias{correct.acrophase.msf}
\title{Get the correct value of acrophase}
\usage{
correct.acrophase.msf(b_rrr, b_sss)
}
\arguments{
\item{b_rrr}{is the beta coefficient for x (beta), which is cos(2πt/τ), τ is period}

\item{b_sss}{is the beta coefficient for z (gamma), which is sin(2πt/τ), τ is period}
}
\value{
correct value of acrophase
}
\description{
Given beta (b_rrr) and gamma (b_sss), get the correct value of acrophase.
}
\examples{

}
