% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.boot.FUN.cont.R
\name{create.boot.FUN.cont}
\alias{create.boot.FUN.cont}
\title{Bootstrap to get standard error of MESOR, amplitude and acrophase contrasts}
\usage{
create.boot.FUN.cont(contrast.frm)
}
\arguments{
\item{contrast.frm}{a string formula specifying the names of the predictors over which emmeans are desired.}
}
\value{

}
\description{
Bootstrap function takes in a fitted merMod object as input and returns the statistics (e.g. standard error) of pair-wise contrasts
of MESOR, amplitude and acrophase
}
\examples{

f1.b<-lmer(hrv~age+gender+T0toT14+Tneg7toT7+
           gender*rrr+gender*sss+
           T0toT14*rrr+T0toT14*sss+
           Tneg7toT7*rrr+Tneg7toT7*sss+
           (1|participant_id),
           data=db.model, na.action = na.omit)

boot.cont<-bootMer(f1.b,FUN = create.boot.FUN.cont(contrast.frm='~T0toT14+Tneg7toT7'), nsim =500, parallel = "multicore", ncpus=8)



}
