% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.boot.FUN.mean.R
\name{create.boot.FUN.mean}
\alias{create.boot.FUN.mean}
\title{Bootstrap to get standard error of mean MESOR, amplitude and acrophase}
\usage{
create.boot.FUN.mean(contrast.frm)
}
\arguments{
\item{contrast.frm}{a string formula specifying the names of the predictors over which emmeans are desired.}
}
\description{
Bootstrap function takes in a fitted merMod object as input and returns the statistics (e.g. standard error) of mean
MESOR, amplitude and acrophase in each group.
}
\examples{
f1.b<-lmer(hrv~age+gender+T0toT14+Tneg7toT7+
           gender*rrr+gender*sss+
           T0toT14*rrr+T0toT14*sss+
           Tneg7toT7*rrr+Tneg7toT7*sss+
           (1|participant_id),
           data=db.model, na.action = na.omit)

boot.mean<-bootMer(f1.b,FUN = create.boot.FUN.mean(contrast.frm='~T0toT14+Tneg7toT7'), nsim =500, parallel = "multicore", ncpus=8)


}
