% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.cosinor.mixed.R
\name{fit.cosinor.mixed}
\alias{fit.cosinor.mixed}
\title{Fit a cosinor mixed model}
\usage{
fit.cosinor.mixed(
  y,
  x = NULL,
  random,
  interaction = NULL,
  data,
  na.action = na.omit
)
}
\arguments{
\item{y}{a vector specifying the outcome variable /phenotype variable}

\item{x}{a vector specifying the predictor variables}

\item{random}{a vector specifying the random effects as specified in lmer formula. e.g. "1|participant_id".}

\item{interaction}{a vector  specifying the variables for interactions}

\item{data}{a dataframe with the x,y, covariates, random variables}

\item{na.action}{how to deal with missing data, default is na.action = na.omit}
}
\description{
Given an outcome, predictor variable, fit the cosinor mixed model with optional covariate effects
}
\examples{

library(lme4)
data("db.model")

## examine the effects of gender on MESOR, acrophase, amplitude on hrv
f<-fit.cosinor.mixed(y="hrv", x="gender", random="1|participant_id", data=db.model)
summary(f)

## examine the interaction effects of gender and baseline bmi categories on MESOR, acrophase, amplitude on hrv
f2<-fit.cosinor.mixed(y="hrv",interaction=c("gender","bmi_baseline_cat"),random="1|participant_id", data=db.model)
summary(f2)

}
