% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromAcrophaseToTime.R
\name{fromAcrophaseToTime}
\alias{fromAcrophaseToTime}
\title{Get the Acrophase and PeakTime}
\usage{
fromAcrophaseToTime(acrophase, period = 24)
}
\arguments{
\item{acrophase}{}

\item{period}{= 24}
}
\description{
Get the Acrophase and PeakTime
}
