% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.contrasts.ci.cosinor.R
\name{get.contrasts.ci.cosinor}
\alias{get.contrasts.ci.cosinor}
\title{Get contrasts and confidence interval for MESOR, Amplitude, Acrophase}
\usage{
get.contrasts.ci.cosinor(
  fit,
  contrast.frm,
  nsim = 500,
  parallel = "multicore",
  ncpus = 8,
  conftype = "norm",
  conflevel = 0.95,
  ...
)
}
\arguments{
\item{fit}{the object from lmer()}

\item{contrast.frm}{a string formula specifying the names of the predictors over which emmeans are desired.}

\item{nsim}{the number of simulations, positive integer; the bootstrap B (or R) for bootMer function.Default is 500.}

\item{parallel}{the type of parallel operation to be used (if any) for bootMer function. The default is "multicore".}

\item{ncpus}{integer: number of processes to be used in parallel operation: typically one would choose this to be the number of available CPUs.Default is 8.}

\item{conftype}{a character string representing the type of interval required.The value must be one of "norm", "basic","perc".}

\item{conflevel}{The confidence level required, default is 0.95.}

\item{...}{additional argument(s) for methods.}
}
\description{
The function takes in a fitted merMod object as input and returns the statistics (e.g. standard error) of pair-wise contrasts
of MESOR, amplitude and acrophase. The output are bootstrapped estimates, 95% confidence intervals and p values for each comtrast
for MESOR, Amplitude and acrophase.
}
\examples{
library(lme4)
library(emmeans)
data("db.model")
f<-fit.cosinor.mixed(y="hrv", x="gender", random="1|participant_id", data=db.model)
summary(f)
get.contrasts.ci.cosinor(f,contrast.frm="~gender")

}
