% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.means.ci.cosinor.R
\name{get.means.ci.cosinor}
\alias{get.means.ci.cosinor}
\title{Get means and confidence interval for MESOR, Amplitude, Acrophase}
\usage{
get.means.ci.cosinor(
  fit,
  contrast.frm,
  nsim = 500,
  parallel = "multicore",
  ncpus = 8,
  conftype = "perc",
  conflevel = 0.95,
  ...
)
}
\arguments{
\item{fit}{the object from lmer() function}

\item{contrast.frm}{a string formula specifying the names of the predictors over which emmeans are desired.}

\item{nsim}{the number of simulations, positive integer; the bootstrap B (or R) for bootMer function.Default is 500.}

\item{parallel}{the type of parallel operation to be used (if any) for bootMer function. The default is "multicore".}

\item{ncpus}{integer: number of processes to be used in parallel operation: typically one would choose this to be the number of available CPUs.Default is 8.}

\item{conftype}{A character string representing the type of interval required.The value must be one of "norm", "basic","perc".}

\item{conflevel}{The confidence level required, default is 0.95.}

\item{...}{additional argument(s) for methods.}
}
\description{
The function will takes in a fitted merMod object, transform the linear/raw coefficients to COSINOR non-linear/transformed coefficients
for MESOR, Amplitude and acrophase and estimated the standard error through bootstrapping to provide the
confidence interval.
}
\examples{

f<-fit.cosinor.mixed(y="hrv", x="gender", random="1|participant_id", data=db.model)
summary(f)
get.means.ci.cosinor(fit=f, contrast.mean.frm="~gender")

}
