% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.pairwise.diff.R
\name{get.pairwise.diff}
\alias{get.pairwise.diff}
\title{Get pairwise difference between two comparison groups}
\usage{
get.pairwise.diff(x)
}
\arguments{
\item{x}{is numeric values of two groups e.g. MESOR between two groups (from just.get.means.cosinor())}
}
\value{
The difference between the two groups. e.g  "MESOR_A_vs_B" and it returns the value of B-A.
}
\description{
Get pairwise difference between two comparison groups
}
