% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/just.get.contrasts.cosinor.R
\name{just.get.contrasts.cosinor}
\alias{just.get.contrasts.cosinor}
\title{Get the pairwise contrasts for MESOR, Amplitude, Acrophase}
\usage{
just.get.contrasts.cosinor(fit, contrast.frm, output = "pairwise.diff", ...)
}
\arguments{
\item{fit}{the object from lmer()}

\item{contrast.frm}{a string formula specifying the names of the predictors over which emmeans are desired.}

\item{output}{output = 'pairwise.diff'.}

\item{...}{other arguments passed on to methods}
}
\value{

}
\description{
Get the estimated marginal means for the specified factors or factor combinations in a linear model,
and transform to non-linear parameters MESOR, amplitude, and acrophase, and get the pairwise contrasts
for MESOR, amplitude and acrophase by the specified factors.
}
\examples{
f1.a<-lmer(hrv~age+gender+
           gender*rrr+
           gender*sss+(1|participant_id),
           data=db.model, na.action = na.omit)

just.get.contrasts.cosinor(fit=f1.a, contrast.frm='~gender')


}
