% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/just.get.means.cosinor.R
\name{just.get.means.cosinor}
\alias{just.get.means.cosinor}
\title{Get MESOR(M), Amplitude(A), Acrophase}
\usage{
just.get.means.cosinor(fit, contrast.frm, ...)
}
\arguments{
\item{fit}{the lmer object from lmer()}

\item{contrast.frm}{a string formula specifying the names of the predictors over which emmeans are desired.}

\item{...}{other arguments passed on to methods}
}
\description{
Get the estimated marginal means for the specified factors or factor combinations in a linear model,
and transform to non-linear parameters MESOR, amplitude, and acrophase.
}
\examples{
 f1.a<-lmer(hrv~age+gender+T0toT14+
            gender*rrr+gender*sss+
            T0toT14*rrr+T0toT14*sss+
            (1|participant_id),
            data=db.model, na.action = na.omit)

 just.get.means.cosinor(fit=f1.a, contrast.frm='~T0toT14')


}
