% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.test.stars.R
\name{stat.test.stars}
\alias{stat.test.stars}
\title{Get the stars for statistical test plots}
\usage{
stat.test.stars(db.delta, contrast.frm)
}
\arguments{
\item{db.delta}{the output from function get.contrasts.ci.cosinor()}

\item{contrast.frm}{a string formula specifying the names of the predictors over which emmeans are desired.}
}
\value{

}
\description{
Get the stars for statistical test plots
}
\examples{
stat.test.stars(db.delta=db.delta, contrast.frm = "~gender")

}
